refactoring registered stuff into modules
