+++ src/main/resources/assets/src/components/summary.js
@@ -1,0 +1,11 @@
+const summary = {
+    templateUrl : 'html/components/summary.html',
+    bindings : {
+        query : '<'
+    },
+    controller: function($scope, $state, $stateParams) {
+        var $ctrl = this;
+    }
+};
+
+export default summary;
