+++ src/main/resources/assets/src/services/types.js
@@ -1,0 +1,15 @@
+function types() {
+    return {
+        SPANS : Types.SPANS,
+        LABEL : Types.LABEL,
+        STRING : Types.STRING,
+        LIST : Types.LIST,
+        LONG : Types.LONG,
+        DOUBLE : Types.DOUBLE,
+        BOOLEAN : Types.BOOLEAN,
+        DATE : Types.DATE,
+        DATE2 : Types.DATE2
+    }
+}
+
+export default types;
